my-flask-neo4j/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
neo4j==5.28.2
python-dotenv==1.1.1
pytz==2025.2
Werkzeug==3.1.3
